requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
brotli==1.1.0
jinja2==3.1.4; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
python-dotenv==1.0.1; python_version >= '3.8'
soupsieve==2.6; python_version >= '3.8'
